View bill, update customer

+++ GemstonesBusinessManagementSystem/GemstonesBusinessManagementSystem/Models/MembershipsType.cs
@@ -10,17 +10,17 @@ namespace GemstonesBusinessManagementSystem.Models
     {
         private int idMembershipsType;
         private string membership;
-        private double target;
+        private long target;
 
         public int IdMembershipsType { get => idMembershipsType; set => idMembershipsType = value; }
         public string Membership { get => membership; set => membership = value; }
-        public double Target { get => target; set => target = value; }
+        public long Target { get => target; set => target = value; }
 
         public MembershipsType()
         {
 
         }
-        public MembershipsType(int idMembershipsType, string membership, double target)
+        public MembershipsType(int idMembershipsType, string membership, long target)
         {
             this.idMembershipsType = idMembershipsType;
             this.membership = membership;
